fix: rename components
