+++ src/pages/BoardView.js
@@ -1,12 +1,12 @@
 import React, { useEffect, useReducer, useState} from "react";
-import Cardreducer from "./BoardReducer";
-import DnDContext from "../components/DragDropContext";
+import BoardReducer from "../reducers/BoardReducer";
+import Board from "../components/Board";
 import _ from "lodash";
 
 export const MyContext = React.createContext();
 
-const Board = () => {
-  const [columns, dispatch] = useReducer(Cardreducer, {});
+const BoardView = () => {
+  const [columns, dispatch] = useReducer(BoardReducer, {});
   const [useData, setUseData] = useState({});
   const [boardVal, setBoardVal] = useState();
 
@@ -50,10 +50,10 @@ const Board = () => {
   return (
     <>
       <MyContext.Provider value={{ columns, dispatch, boardVal }}>
-        <DnDContext></DnDContext>
+        <Board></Board>
       </MyContext.Provider>
     </>
   );
 }
 
-export default Board;
+export default BoardView;
